Add files via upload
Open Corner has conducted a seminar on #IoT to enhance the conceptual and practical knowledge of attendees. This seminar was given by Bhavik Prajapati, Ankit Rana, and Tanuka Roychowdhury. All our speakers have 2+ years of industrial experience. They gave an excellent explanation of what is IoT and how to implement it for our routine life. IoT device and software was given to every attendee to get a better understanding.